Split native ABI manifest parsing
